Some code reorginazation and comments
